+++ KeyboardTester/KeyboardTester.sln
@@ -3,14 +3,14 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.7.34031.279
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "KeyboardTester", "KeyboardTester\KeyboardTester.csproj", "{2DB5E589-9D24-4FAE-8E51-3C6914E824CF}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "KeyboardTesterApp", "KeyboardTesterApp\KeyboardTesterApp.csproj", "{2DB5E589-9D24-4FAE-8E51-3C6914E824CF}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "UnitTests", "UnitTests\UnitTests.csproj", "{59C6DB2D-9AA1-4465-8594-C03B716375B3}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{5A697254-F91D-4125-B198-8CAFBBB8F1E0}"
 	ProjectSection(SolutionItems) = preProject
 		.editorconfig = .editorconfig
-		Dockerfile_lint. = Dockerfile_lint.
+		Dockerfile_lint = Dockerfile_lint
 	EndProjectSection
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "UITests", "UITests\UITests.csproj", "{85D3D854-E555-41BB-9D01-B57152498554}"
