--- KeyboardTester/KeyboardTester/GlobalUsings.cs
@@ -1,10 +1,0 @@
-﻿global using KeyboardTester.Models;
-global using KeyboardTester.Models.DataTransferObjects;
-global using KeyboardTester.Models.KeyboardLayouts;
-global using KeyboardTester.Services;
-global using KeyboardTester.Util;
-global using Newtonsoft.Json;
-global using System.Diagnostics;
-global using System.Runtime.InteropServices;
-global using System.Text.RegularExpressions;
-global using static KeyboardTester.KeyboardHook;
