+++ KeyboardTester/KeyboardTesterApp/Dockerfile_publish
@@ -1,0 +1,3 @@
+FROM mcr.microsoft.com/dotnet/sdk:6.0-windowsservercore-ltsc2019
+WORKDIR /app
+ENTRYPOINT ["dotnet", "publish", "KeyboardTesterApp/KeyboardTesterApp.csproj", "--self-contained", "--runtime", "win-x64", "--property", "WarningLevel=0", "--configuration", "Release", "--output", "artifacts/"]
