+++ KeyboardTester/KeyboardTesterApp/KeyboardTesterForm.resx
