+++ KeyboardTester/KeyboardTesterApp/Models/Key.cs
@@ -1,4 +1,4 @@
-﻿namespace KeyboardTester.Models
+﻿namespace KeyboardTesterApp.Models
 {
     public class Key : Button
     {
