+++ KeyboardTester/KeyboardTesterApp/Models/KeyboardLayouts/ToughbookKeyboardLayout.cs
@@ -1,4 +1,4 @@
-﻿namespace KeyboardTester.Models.KeyboardLayouts
+﻿namespace KeyboardTesterApp.Models.KeyboardLayouts
 {
     public class ToughbookKeyboardLayout : KeyboardLayout
     {
