+++ KeyboardTester/KeyboardTesterApp/Services/InformationBoxService.cs
@@ -1,4 +1,4 @@
-﻿namespace KeyboardTester.Services
+﻿namespace KeyboardTesterApp.Services
 {
     public class InformationBoxService
     {
