+++ KeyboardTester/KeyboardTesterApp/Services/KeyboardLayoutService.cs
@@ -1,4 +1,4 @@
-﻿namespace KeyboardTester.Services
+﻿namespace KeyboardTesterApp.Services
 {
     public class KeyboardLayoutService
     {
