+++ KeyboardTester/UITests/GlobalUsings.cs
@@ -1,9 +1,9 @@
-global using KeyboardTester.Models.DataTransferObjects;
+global using KeyboardTesterApp.Models.DataTransferObjects;
+global using KeyboardTesterApp.Util;
 global using Microsoft.VisualStudio.TestTools.UnitTesting;
 global using Newtonsoft.Json;
 global using OpenQA.Selenium.Appium;
 global using OpenQA.Selenium.Appium.Windows;
-global using System.Drawing;
 global using System.Runtime.InteropServices;
 global using System.Windows.Forms;
 global using static UITests.InputSender;
