+++ KeyboardTester/UITests/UITests.csproj
@@ -29,7 +29,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\KeyboardTester\KeyboardTester.csproj" />
+    <ProjectReference Include="..\KeyboardTesterApp\KeyboardTesterApp.csproj" />
   </ItemGroup>
 
 </Project>
