+++ KeyboardTester/UnitTests/Dockerfile_test
@@ -1,3 +1,3 @@
 FROM mcr.microsoft.com/dotnet/sdk:6.0-windowsservercore-ltsc2019
 WORKDIR /app
-ENTRYPOINT ["dotnet", "test", "UnitTests/UnitTests.csproj", "--logger", "html;logfilename=TestResults.html"]
+ENTRYPOINT ["dotnet", "test", "UnitTests/UnitTests.csproj", "--logger", "html"]
