+++ KeyboardTester/UnitTests/GlobalUsings.cs
@@ -1,7 +1,6 @@
-global using KeyboardTester;
-global using KeyboardTester.Services;
-global using KeyboardTester.Util;
+global using KeyboardTesterApp;
+global using KeyboardTesterApp.Services;
+global using KeyboardTesterApp.Util;
 global using Microsoft.VisualStudio.TestTools.UnitTesting;
-global using System.Drawing;
 global using System.Windows.Forms;
-global using static KeyboardTester.KeyboardHook;
+global using static KeyboardTesterApp.KeyboardHook;
